#462 Extract MUI Icons mock to separate file for cleanliness
